The older (I think) lancelot shortcode plugin
At some point the org-mode shortcode changed and either started or stopped
passing in `post` as an argument and instead started passing in a dictionary
with `post` as a key. I think that the dictionary is the newer form of behavior.
This version of lancelot has the old behavior, which some of my repositories
still use.